requirements.txt
Flask==2.0.1
requests==2.25.1
werkzeug==2.0.3
python-dotenv==0.21.0 